CASMINST-2355: update mountain/hill routes on NCNs
This script updates the cloud-init user-data in bss for all NCNs.
The update is comprised of "write_files" stanzas that will create
ifroute files on NCNs when they boot into the 1.5 image.

The logic for querying bss is lifted from a previous WAR for this
issue. Once the routes are determined, they are written out to a
a json file with cloud-init user-data. This file is then passed
into csi for updating bss.