CASMISNT-1629 CASMINST-1821: clarify cabinets.yaml advice
Contrary to popular documented opinion, csi does not actually
merge cabinet-related data from the cli and a cabinets.yaml file
into a coherent set of data.  If a cabinets.yaml file is used, csi
ignores all related cabinet command-line flags.

csi has been modified to emit a warning when the user attempts this.
In a future release it will likely be a hard error.