SAT-739: Add backup/restore of SAT local files to upgrade script
This adds two stages:
* BACKUP_SAT_LOCAL_FILES
* RESTORE_SAT_LOCAL_FILES

The first handles copying files off the specified 'upgrade' NCN,
and the second restores them.

Test Description:
I was not able to test this full script, but I ran a script consisting
of the first stage followed by the second stage with a pause in between,
showing that the stages could copy the files off the NCN and restore
from the backup.