CRAYSAT-941: Update Ceph instructions in K8s cluster start procedure
The `sat bootsys boot --stage platform-services` command has been
updated to unfreeze Ceph if it is frozen and then wait for it to reach a
healthy state. Update the Ceph instructions in the procedure for
starting the Kubernetes management cluster as follows:

* Remove manual steps for unfreezing Ceph
* Update the instructions and example output for the case when Ceph
  fails to reach a healthy state.
* Add instructions to re-run the `sat bootsys boot --stage
  platform-services` after Ceph is healthy, if necessary.

Test Description:
Starting platform services with this `sat bootsys` command was tested on
a virtual Shasta system, and example output was gathered from that
environment.