DOCS: Added script to set SNMP creds for mgmt leaf switches.
Added script to set SNMP creds for mgmt leaf switches.
This makes it easy for an admin to set the default SNMP
user and passwords for the leaf switches without having
to do it manually.  It uses 'expect' to interact with
the switches.

This change is for the docs, to include the usage of this
new script.

REMINDER: HAVE YOU INCREMENTED VERSION NUMBERS? E.G., .spec, Chart.yaml N/A

* Resolves CASMHMS-5183

N/A

No risk, docs change only.