CASMINST-3508 Rename "Redeploy PIT Node"
The stage for "Redeploying the PIT Node" is confusing contextually
because it does not actually result in a redeploymnet of a PIT. The end
result is a deployed NCN, specifically the final NCN in the CSM install.

This change refactors cases of "Redeploy PIT Node" to "Deploy Final NCN"
to express this clarity, emphasizing that were still working on a unique
node (the node running the LiveCD is the final NCN) and need special
procedure to wrap-up the install of the NCNs.