CRAYSAT-1345: Remove unnecessary "| tee sat.capture-state"
This change fixes a command line instruction which incorrectly told the
user to tee the contents of stdout to the file `sat.capture-state`. When
running `sat bootsys shutdown --stage capture-state`, no information is
expected to be written to stdout, and the state capture is instead
written to S3. Thus the `tee` pipeline should be removed as not to
confuse the admin.