CASMTRIAGE-2988: Instructions on adding xnames to BOS session templates
When a new blade is added to the system, the sysadmin will want to boot
these new nodes. This docuentation update checks if a BOS Session
template uses the node_list to specify names explicitly. If so, the
new nodes need to be added into the Session template, so they can be
booted using that template.