CASMINST-4149: never wipe USB devices
The wipe commands don't care if a device is mounted or not. They
will wipe it regardless.

Modify the code so the list of devices to wipe doesn't include the
USB. This borrows code that is in dracut for the same reason.

While here, reduce the number of lines needed to figure out which
device is the USB device. (IMO, this code doesn't need to care about
whether the USB device is mounted or not, but it's a smaller code
change to leave that logic in place. And it's harmless.)