MTL-1656 Move NCN Background Info
The technical information for the NCN needs to decouple from the
high-traffic repository for installs and upgrades. Not only is the
information not relevant for every user of an install and upgrade, it is
at risk of being neglected. The context for NCN technical and
installs/upgrades is different. Technical is not always necessary to
read, and may be read at a later point if not an earlier point. It may
be useful to refer to technical documents during an install/upgrade
operation, but whether a user follows the reference or not is up to
them. Therefore these are being moved out of this repo into a better
place.