CASMHMS-5522: NCN Remove changed stop REDS
Changed the documenation to have the user shutdown cray-reds before
doing the remove. And added a note that the BSS restart may timeout if
there are only 2 ncn workers.

There was a race condition where cray-reds would add back a component
to HSM while the remove_management_ncn.py was removing the component
from BSS, SLS, and HSM.

Added a default ipmi username to ncn_status.py. This makes the script
match the documentation.