CASMTRIAGE-3536: time hack (#1869)
This commit makes the following changes:

1. In upgrade/prerequisites.sh, correct/verify the chrony config
for every node. Do this instead of doing it one node at a time
as they are upgraded. After `csm_ntp.py` has been run on a node,
use `chronyc makestep` to force time to sync immediately. With
this approach, time should be in sync across the cluster prior
to the upgrade beginning in earnest.

2. Once an upgraded node is online and settled, use `chronyc makestep`
to force time to sync immediately.