CASMINST-4833 Updates (#1911)
Git-tags are stable builds that should publish to a sub-directory
containing their major and minor versions (e.g. v1.2.6 pushes to
docs-csm/1.2).

Every git-tag should also push a latest RPM. This means there is a risk
that rebuilding a previous tag will overwrite the latest RPM, this can
be remedied by rebuilding the latest tag for the CSM release.