CASMINST-4863: run csm_ntp.py on a newly imaged/upgraded node
CSM-1.2 images contain a pool.conf file that needs to be removed.
This is one of the things csm_ntp.py can fix, so invoke it before
attempting to sync the clock.

Also add a wait loop like we have in prerequsites.sh.  The 5s wait
is not always enough.

This is ugly from a code duplication POV, but less risky than trying
to reorganize the code at this point in time.

(cherry picked from commit 0fe6844775050a79caaf6d6e01315e4fb5da5566)