CASMINST-4684 Remove Manual Wipe Steps (#1878)
* CASMINST-4684 Remove Manual Wipe Steps

With the fix from MTL-1288, manually wiping is no longer necessary. To
rebuild an NCN after following respective procedures for expunging the
NCN from its clusters, only metal.no-wipe=0 must be set and the node PXE
booted.

* CASMINST-4684 Update rebuild scripts

These rebuild scripts live in a 1.2 directory but will be used in 1.3.
I am updating these so that the removal of the wipe is included whenever
these scripts become used in 1.3.