NETETH-1987 CASMINST-4877: blacklist the `rpcrdma` kernel module (#1952)
NETETH-1987 CASMINST-4877: blacklist the rpcrdma kernel module

In order to not cause problems during the Slingshot Host Software
install we need to blacklist the rpcrdma kernel module.  This is
accomplished by added a kernel boot param to blacklist the module.

The doc change instructs the admin to run this script prior to
starting the upgrade.  This ensures that the worker nodes have
this module blacklisted on their first boot.  This also ensures
the boot params, including this new parameter, get automagically
added to the grub.cfg during the first network boot.