CASMINST-4915 ncn-image-modification
This adds the ncn-image-modification.sh script call to the
pre-installation.md page, that is necessary to invoke before
invoking the preflight tests.

The call is added right after the images are extracted.

Additionally the step to extract the images was changed to extract
images into a versioned directory by using a faster rsync command. The
versioned directory will help keep things neat, and will align better
with NCN smoke-testing/metal-testing.

The faster rsync command entails using hard-links, which avoid wasting
space by avoiding duplication of the NCN artifacts on the same file
system.