CASMTRIAGE-3665: Patch sma kafka cluster resource and add listeners
During this part of the upgrade, if the kafka sma cluster object does not have
listeners setup correctly, the upgrade cannot continue.

So detect if/when the listeners are not setup yet, and if so json merge patch
the data in.

Additionally ensure that the kafka-kafka pods are at a Running status and Ready
state of 1/1. Old installs have two containers and we want to ensure we're
restarting the new setup at this point. So if there are any non Running pods or
pods not at 1/1 Ready, delete the pods as well.