CASMINST-4948: Fix broken link in `upgrade/README.md`
The file `upgrade/1.2/README.md` moved to `upgrade/upgrade_ncn_nodes.md`
in CASMINST-4943, but the link in `upgrade/README.md` was not updated,
so the link was broken. Update the link. Note that the contents of that
file still needs to be updated for CSM 1.3. That is not done here.

Test Description:
Checked that link was no longer broken when viewing this branch in
GitHub.