CASMINST-4948: Use CSM 1.3.0 version in example
Update the version of CSM shown in example output from querying the
product catalog from 1.2.0 to 1.3.0. Also update the import_branch in
that example to show the current 1.10.0 version. This may change before
the release of CSM 1.3.0, but it's okay because it's noted that the
output is an example.

Update the value of the `--csm-release` option for CSM 1.3.0 in the call
to `apply_csm_configuration.sh`. This will default to the latest
version, but if the user specifies it here, they would want to specify
CSM 1.3.0.

Test Description:
Viewed rendered docs on GitHub.com.