CASMINST-4948: Update CSM version in NCN Personalization Steps (#2048)
* CASMINST-4948: Fix broken link in `upgrade/README.md`

The file `upgrade/1.2/README.md` moved to `upgrade/upgrade_ncn_nodes.md`
in CASMINST-4943, but the link in `upgrade/README.md` was not updated,
so the link was broken. Update the link. Note that the contents of that
file still needs to be updated for CSM 1.3. That is not done here.

Test Description:
Checked that link was no longer broken when viewing this branch in
GitHub.

* CASMINST-4948: Use CSM 1.3.0 version in example

Update the version of CSM shown in example output from querying the
product catalog from 1.2.0 to 1.3.0. Also update the import_branch in
that example to show the current 1.10.0 version. This may change before
the release of CSM 1.3.0, but it's okay because it's noted that the
output is an example.

Update the value of the `--csm-release` option for CSM 1.3.0 in the call
to `apply_csm_configuration.sh`. This will default to the latest
version, but if the user specifies it here, they would want to specify
CSM 1.3.0.

Test Description:
Viewed rendered docs on GitHub.com.