CASMINST-5041 Remove Dupe Steps and Fix Password
This fixes the Change NCN Password and SSH Keys on PIT Node page to use
the correct paths to the images.

This also removes the duplicate steps for invoking
`ncn-image-modification.sh` and `set-sqfs-links.sh`, which were invoked
in both the `pre-installation.md` and `deploy_non-compute_nodes.md`
pages. The steps in `deploy_non-compute_nodes.md` were redundant, since
`pre-installation.md` is _required_ before starting the
`deploy_non-compute_nodes.md` page.

This also fixes a typo that went unnoticed, the storage-ceph image path
was using the `kubernetes_version` variable. This wasn't causing
problems because the two versions are _usually_ the same, but since they
potentially can differ this ensures that either case is handled.