CAST-30583: Kafka restart bugfixes
Noticed the podsready() function had a bit of a no-op so removed the field
selector in that kubectl call.

Comment of retrying for 300 seconds only was invalid so added some logic to
ensure the restart script only tries for 5 minutes to match.

And typo fixes for the general script as well where my prior testing appeared to
miss issues.