CASMINST-4462: NCN image modification procedure  (#2100)
* CASMINST-4462 CASMINST-4805: NCN image modification procedure (#1506)

CASMINST-4462: NCN image modification procedure

Create a procedure for using CFS to modify NCN boot artifacts prior
to booting them. This procedure mostly guides the reader through
existing documentation, with only a few new steps added that are
specific to interacting with NCN artifacts.

* Adjust the wipe step

A full manual wipe is not needed in CSM-1.3.  However, we do
need to remove the bootloader so that the node will accept a
new squashfs with the same name. We also need to ensure that
`metal.no-wipe` is set to `0`.