CASMTRIAGE-3993 Remove invalid `csi` call (#2314)
* CASMTRIAGE-3993 Remove invalid `csi` call

`csi handoff upload-utils` no longer exists, and the call itself was no
longer necessary. Now that the function no longer exists this call
fails, but since it is no longer necessary this call can be removed.

* Fix idempotency and double-return

The `KUBERNETES_VERSION` and `CEPH_VERSION` vars contain newlines on
re-runs, since the `secure` and non-secure images are found by the
`find` commands.

This fixes some idempotency issues with the `UPLOAD_NEW_NCN_IMAGE` step
as well, preventing re-runs of `ncn-image-modification.sh` if it was ran
already.