CASMHMS-5462: Add and update procedures to add river cabinets to a system.
CASMHMS-5462: Added a procedure to configure the root user account a HPE
Node. Updated the Gigabyte controller procedure.

Added a procedure to enable IPMI access on new HPE nodes.

Reworked the add NCN procedure to make use of the enable IPMI access, and configure root user procedures at a time where the BMC is actually reachable.

Also added a procedure to set time on HPE nodes. This allows the docmentation
to point toward documentation that is valid after the system is
installed, instead of on the PIT node.

Updated the add a standard rack node to reference WLM documentation.