CASMTRIAGE-4039 Fix disk selection
Disk selection was inadvertently including `sr0` (remoteISO) devices and
USBs. The USBs were assumed to not be plugged in, and the `sr0` devices
were assumed to never be _first_. The loki install today proved that
`sr0` does come first on occaision, and it proved that admins leave USB
sticks in.

This change omits USB and ROM devices, and ensures that the `disk`
variable only contains *1* item.

(cherry picked from commit d49679803185643923ae9fa8e5328ff93524135d)