CASMINST-3902: Add and update procedures related to adding river cabinets to a system. (#2345)
* CASMHMS-5462: Add and update procedures to add river cabinets to a
system.

CASMHMS-5462: Added a procedure to configure the root user account a HPE
Node. Updated the Gigabyte controller procedure.

Added a procedure to enable IPMI access on new HPE nodes.

Reworked the add NCN procedure to make use of the enable IPMI access, and configure root user procedures at a time where the BMC is actually reachable.

Also added a procedure to set time on HPE nodes. This allows the docmentation
to point toward documentation that is valid after the system is
installed, instead of on the PIT node.

Updated the add a standard rack node to reference WLM documentation.

* Apease the spell checker

* Apply suggestions from code review

Co-authored-by: Mitch Harding <mitchell.harding@hpe.com>
Signed-off-by: Ryan Sjostrand <77114568+rsjostrand-hpe@users.noreply.github.com>

* Apply suggestions from code review

Co-authored-by: Mitch Harding <mitchell.harding@hpe.com>
Signed-off-by: Ryan Sjostrand <77114568+rsjostrand-hpe@users.noreply.github.com>

* Apply suggestions from code review

* Appease the markdown linter

* Change wording in the Update HPE Node BIOS time procedure

* Apply suggestions from code review

* Apply suggestions from code review

* Apply suggestions from code review - Rework Add Root Service Account for Gigabyte Controllers to flow betterZ

* Apply suggestions from code review - Move steps to view SNMP redentials into a new section

* Apply suggestions from code review

* Apply suggestions from code review

* Apply suggestions from code review

Co-authored-by: Mitch Harding <mitchell.harding@hpe.com>
Signed-off-by: Ryan Sjostrand <77114568+rsjostrand-hpe@users.noreply.github.com>

Signed-off-by: Ryan Sjostrand <77114568+rsjostrand-hpe@users.noreply.github.com>
Co-authored-by: Mitch Harding <mitchell.harding@hpe.com>