Fix `KVER` usage
The `-default` suffix needs to be stripped off when referring to the
package version. The package is also called `kernel-default-debuginfo`
not `kernel-default-debug`.

The `--plus-content debug` may or may not be needed but it doesn't hurt.

(cherry picked from commit bbb1a9af5bd85f35f8b6e52b7265235c86cdb524)