CASMPET-6002: Add workaround for mounting boot-images on workers
When the worker nodes boot, they try to mount the boot-images bucket as
an s3fs mount. However, the Ceph cluster is not unfrozen yet, so this
mount will fail. After `sat bootsys boot --stage platform-services` is
executed, Ceph will be unfrozen, and at this point, it will be possible
to mount the `boot-images` filesystem. This filesystem is needed to
allow the cray-cps-cm pods to start successfully.

Add a manual workaround to mount this filesystem, along with any other
file systems of type s3fs on worker nodes.

Test Description:
This is what Don said he did in CASMTRIAGE-4144 to get the CPS pods to
come up.

This particular command has been tested on fanta as well. It has been
tested with multiple sequential invocations to ensure it is idempotent
as well.

(cherry picked from commit 20b72ca5780de1c890a4fe2cc36589aaef2eed34)