CASMPET-6002: Document Ceph workarounds for System Power On Procedures (#2508)
* CASMPET-6002: Add information about Ceph deep scrubbing

Add information about Ceph deep scrubbing and how to proceed when the
Ceph cluster is in a `HEALTH_WARN` state during the system power on
procedures.

Test Description:
Ran command to view Ceph deep scrub schedule on a system to ensure it
worked.

* CASMPET-6002: Add workaround for mounting boot-images on workers

When the worker nodes boot, they try to mount the boot-images bucket as
an s3fs mount. However, the Ceph cluster is not unfrozen yet, so this
mount will fail. After `sat bootsys boot --stage platform-services` is
executed, Ceph will be unfrozen, and at this point, it will be possible
to mount the `boot-images` filesystem. This filesystem is needed to
allow the cray-cps-cm pods to start successfully.

Add a manual workaround to mount this filesystem, along with any other
file systems of type s3fs on worker nodes.

Test Description:
This is what Don said he did in CASMTRIAGE-4144 to get the CPS pods to
come up.

This particular command has been tested on fanta as well. It has been
tested with multiple sequential invocations to ensure it is idempotent
as well.

* CASMPET-6002: Fix a broken link to Management Node Image Customization