CASMPET-6004: Add validation for checking the user input against etcd
This adds the ability for encryption.sh --enable/disable --aesgcm ...
to validate the user input makes logical sense against what is found in the etcd
database.

What this means is essentially if a user tries to setup encryption with a key
that isn't found in etcd, the script will reject the update as being logically
invalid.

This is to ensure that kubernetes secrets can always be read given a specific
encryptionn configuration.

Also added switches --status to display the status of encryption.

And added a --restart switch to make restarting the daemonset easier to force updates.