CASMINST-5425: Remove "upgrade" language
This commit removes the "upgrade" language from the following two
procedures, since they are used in the context of both upgrades
and fresh installs:
* Worker Image Customization
* NCN Node Personalization

Additionally, in the case of node personalization, the word "worker"
is removed and the documentation is updated to refer to management NCNs
instead, because that is the correct set of nodes affected by the node
personalization process - not just workers.

Finally, mark the steps to acquire current CPE and Analytics CFS config
values only necessary during an upgrade. In a fresh install, these
configs will have already been installed.

Update links from Stage_0_Prerequisites.md accordingly.