CASMINST-5425: Revise NCN image customization and personalization language to fit fresh install (#2615)
* CASMINST-5425: Remove "upgrade" language

This commit removes the "upgrade" language from the following two
procedures, since they are used in the context of both upgrades
and fresh installs:
* Worker Image Customization
* NCN Node Personalization

Additionally, in the case of node personalization, the word "worker"
is removed and the documentation is updated to refer to management NCNs
instead, because that is the correct set of nodes affected by the node
personalization process - not just workers.

Finally, mark the steps to acquire current CPE and Analytics CFS config
values only necessary during an upgrade. In a fresh install, these
configs will have already been installed.

Update links from Stage_0_Prerequisites.md accordingly.

* CASMINST-5425: Skip creating CFS configuration if using bootprep

This commit modifies the Management Node Image Customization procedure
to make creating the CFS configuration only necessary if not using
bootprep. Also, a small error is fixed in which a configuration was
being referred to as a session in two places.

* CASMINST-5425: Give exact customization CFS command

* Give the exact `cray cfs sessions create` command that would be
  needed in the case of management node image customization.
* Continue to refer to "Create an Image Customization CFS Session"
  page, but only as a reference.
* Remind the reader that they should have set $IMS_IMAGE_ID
  in a previous step.