CASMINST-5434 Run MTL-1919 WAR after NCN Upgrade
This WAR was ran before the upgrade in order to facilitate PXE booting,
but it is now known that it should also run _after_ the upgrade to fix
disk boot entries.

Running this after the upgrade will ensure that disk boot entries will
appear properly in the BIOS boot selection menu.