CRAYSAT-1596 : Fix typos in "Accessing SAT Bootprep Files" (#2644)
CRAYSAT-1596: Fix typos in "Accessing SAT Bootprep Files"

Remove '>' characters that make the `vcs-creds-helper.sh` script
generation incorrect. This commit also fixes a typo in the word
'bootprep', which was misspelled 'bootrep'.

Test Description: I viewed the resulting branch of this page in
github, ran the command which generates the vcs-creds-helper.sh
script, and manually checked that it was created as expected.