CASMINST-5474: Explicit instructions for setting layers
This commit provides more explicit instructions for setting up the CPE
and Analytics layers in NCN personalization in the case of an upgrade,
in which CPE and Analytics have not yet been installed. Prior to this
change, some knowledge was assumed about how to determine the current
playbook, commit hash and product values already in use. This commit
provides example output from `cray cfs components describe` as well
as explicit instructions on how to edit the bootprep file so that the
`playbook`, `url` and `commit` values are correct. It explicitly says
to use the `git` key in the bootprep layer definitions and provides
an example of what those layers should look like, based on the example
output of `cray cfs components describe`.

Finally, add optional steps to delete the temporary files,
management-bootprep-node-personalization.yml and
management-node-image-customizatino.yml, both of which are only needed
temporarily in the context of their respective procedures.
This is an especially good idea in the upgrade context of ncn personalization,
in which this file will likely become stale after the new versions of CPE and
Analytics have eventually been installed.