CASMINST-5470: Variable name and path consistency in IMS Image Import (#2656)
This commit adds variable name consistency to "Import an External Image
to IMS". In the event that the user is following "Management Node Image
Customization," they will have already set the following three
variables:

* IMS_ROOTFS_FILENAME
* IMS_KERNEL_FILENAME
* IMS_INITRD_FILENAME

This commit adds a reference to the "Management Node Image
Customization" procedure in "Import an External Image to IMS", reminding
the user that they may have already set these variables.

The variable is used in a slightly different way between the two
procedures. In "Management Node Image Customization" it is assumed
that the variables directly refer to filenames that exist locally. In
"Import an External Image to IMS", this was not being assumed, and as
a result a second variable called "ROOTFS_FILENAME" was being defined
which referred to the exact path to the file (confusingly, a nearly
simimlar name to IMS_ROOTFS_FILENAME).

This commit changes the procedure such that all three variables
MUST be the names of files that exist in your local working directory.
In the case of uploading the kernel and initrd, this also means that the
`image-root/boot/` directories are no longer specified when uploading
the kernel and initrd via `cray artifacts` - (in the context
of "Management Node Image Customization" these were completely incorrect
anyway).

As a small fix, the steps to calculate the initrd for kernel and
ramdisk are now their own steps instead of included in the code blocks
under the `cray artifacts` commands that upload them.