CASMINST-5552 Enable boot-order WAR for Upgrades from 1.0.X
Systems upgrading from CSM 1.0.X to 1.2.2+ do not have the `lib.sh`
dependency needed by the WAR `metal-lib.sh` library. This adds that
dependency, enabling the WAR to be ran on CSM 1.0.X images.

When this WAR was originally added, it was intended to run on CSM 1.2.0
or CSM 1.2.1 systems upgrading to CSM 1.2.2 - CSM 1.2.X images include
`lib.sh`, whereas CSM 1.0.X images do not have this file.