CASMPET-6111: Update arg parsing of this script to be more strict (#2751)
Before what would happen is if you did:
./script ncn-w001,ncn-w002 ncn-w003

Only ncn-w003 would be updated as it was the last thing to update the vars.

So instead once the first arg with ncn-... is found, consider that the terminal
case and then if more args matching it are found, exit with a warning message to
stdout and do no work.

Ideally we should make this script just accept the following:
./script ncn-w001 ncn-w002 etc...

so we can just have users do
./script $(kubectl get nodes ... some selector)

Instead of forcing callers of the script to join arguments with a comma. That
steps largely unnecessary and error prone.