CASM-3505: Implement Argo WorkflowTemplate for ims-upload
This change adds a WorkflowTemplate which implements the ims-upload
operation for the IUF. It handles setting up secrets and running the
`load-ims-artifacts` image in order to upload artifacts to S3 and create
the IMS image.

Test Description:
Run workflow template with product distributions and mocked-out
global_params, and ensure that IMS artifacts are uploaded successfully.