CRAYSAT-1626: Document workaround for `sat bootsys` bug
Document a workaround for a `sat bootsys` bug that occurs when waiting
for the `hms-discovery` Kubernetes cronjob to be scheduled.

Based on my testing with a reproducer script on shandy, it seems that
the issue occurs with the first time waiting on the hms-discovery
cronjob to be scheduled after it is resumed (suspend set to false).
The last schedule time reported by K8s for the cronjob is unreliable
during that period. Afterwards, it seems to fix itself, so it should be
fine to just run the `sat bootsys` command again.

Also remove the note about specifying `--discovery-timeout 600`. This is
the default already. The 62-second timeout that was seen was a separate
timeout waiting for the hms-discovery cronjob to be scheduled, and that
is computed by looking at the cronjob's schedule and adding a grace
period of 60 seconds.

Test Description:
Tested this approach with a reproducer script on shandy as described in
detail in CRAYSAT-1623.