CASM-3589: Add product catalog update to nexus-setup
This commit adds a step in the nexus-setup operation so that the output
of the cray-nexus-setup image is recorded in the cray-product-catalog.

As a part of this change, this commit changes the workflow to
temporarily use an unstable build of cray-nexus-setup, until CASM-3758
is resolved. This commit also removes a place where the cray-nexus-setup
image was being set in stages.yaml, as this was not actually used.

Test Description: Modified the nexus-setup-template workflow on frigg
and ran the IUF CLI through the deliver_product stage.