DVS-2827: cover fresh install case
There is a moment in a fresh install where CSM (including our hook) is
present but COS (including DVS) is not yet installed.  We can avoid
blocking rebuilds in this moment by checking the target node to see if
DVS is present, and exit 0 before further checks if it is not.
Generally speaking, if DVS is not loaded on a node, it cannot be
contributing to HA.  This approach may also help avoid blocking a
rebuild in various disaster recovery scenarios.