CASM-3588: Change order of catalog updates
This commit moves the catalog update steps immediately after the stages
that produce the output to insert into the catalog. This is to ensure
that the prom-metrics steps occur last, but also so that it is clearer
what steps consume the output of the previous step.

(cherry picked from commit 09818893378ba134f33164c0f9780e9c59fce788)