CASM-3590: Add catalog updates to nexus-helm-upload-template
This commit adds a step into the workflowtemplate that uploads Helm
charts into nexus. This step gathers the output of the previous step
as an input parameter.

As a part of this commit, the cray-nexus-setup image used by the
workflowtemplate is temporarily changed to an unstable build, adding
a TODO comment to update that to a stable build when one is available.

(cherry picked from commit 5754204fe4292cdeef99cb218244c5e921c5bd2d)