CASM-3588 CASM-3590 Add catalog to docker and helm uploads (release 1.4) (#3052)
* CASM-3588: Add product catalog output to nexus-docker-upload

This commit adds an output parameter to nexus-docker-load-template.
Additionally, the image used by both nexus-setup and nexus-docker-upload
is updated to a newer version of cray-nexus-setup, though the TODO item
still remains to use a stable build of it.

(cherry picked from commit 17d732a303400eda86a2fdbbaabc416167128bdd)

* CASM-3588: Change order of catalog updates

This commit moves the catalog update steps immediately after the stages
that produce the output to insert into the catalog. This is to ensure
that the prom-metrics steps occur last, but also so that it is clearer
what steps consume the output of the previous step.

(cherry picked from commit 09818893378ba134f33164c0f9780e9c59fce788)

* CASM-3590: Add catalog updates to nexus-helm-upload-template

This commit adds a step into the workflowtemplate that uploads Helm
charts into nexus. This step gathers the output of the previous step
as an input parameter.

As a part of this commit, the cray-nexus-setup image used by the
workflowtemplate is temporarily changed to an unstable build, adding
a TODO comment to update that to a stable build when one is available.

(cherry picked from commit 5754204fe4292cdeef99cb218244c5e921c5bd2d)

* CASM-3758: Use stable build of cray-nexus-setup

This commit updates all the workflowtemplatest that use cray-nexus-setup
so that a stable build of the image is used.

(cherry picked from commit 73917b7731591fdbf95b1c033da698e57e3cc5f0)