CASM-3588, CASM-3590, CASM-3758: (IUF) add product catalog updates to docker and helm uploads (#3040)
* CASM-3588: Add product catalog output to nexus-docker-upload

This commit adds an output parameter to nexus-docker-load-template.
Additionally, the image used by both nexus-setup and nexus-docker-upload
is updated to a newer version of cray-nexus-setup, though the TODO item
still remains to use a stable build of it.

* CASM-3588: Change order of catalog updates

This commit moves the catalog update steps immediately after the stages
that produce the output to insert into the catalog. This is to ensure
that the prom-metrics steps occur last, but also so that it is clearer
what steps consume the output of the previous step.

* CASM-3590: Add catalog updates to nexus-helm-upload-template

This commit adds a step into the workflowtemplate that uploads Helm
charts into nexus. This step gathers the output of the previous step
as an input parameter.

As a part of this commit, the cray-nexus-setup image used by the
workflowtemplate is temporarily changed to an unstable build, adding
a TODO comment to update that to a stable build when one is available.

* CASM-3758: Use stable build of cray-nexus-setup

This commit updates all the workflowtemplatest that use cray-nexus-setup
so that a stable build of the image is used.