CASMINST-5900: direct cray-product-catalog-update output to /dev/null
When the code to update the cray product catalog with NCN image info was
added, the output for the `podman` call was not redirected to /dev/null.
As a result, that output was processed by processed by `prerequisites.sh`,
which lead to other problems downstream.

This commit does two things:
1. Redirects stdout from the podman call to /dev/null
2. Adds a regex check on the output from the script to ensure the only
thing we're getting is a valid UUID.