CASMINST-5876: Add support for `CF_*_GITEA_REPO` env var (#3173)
This change adds a `cf_update_gitea_repo` parameter to the
`gitea-update-content` step of the `vcs-update-working-branch` workflow
template, as well as to the `gitea-upload-content` step of the
`vcs-upload-content` workflow template. This allows products to specify
a name which does not conform to the `<product>-config-management`
default.

Test Description:
Test `vcs-update` and `vcs-upload` operations of IUF.

Signed-off-by: Jack Stanek <90287236+jack-stanek-hpe@users.noreply.github.com>
Co-authored-by: Ryan Haasken <77809410+haasken-hpe@users.noreply.github.com>
(cherry picked from commit 9ab49d2a42eeb537b325aec1111f0538870477ba)