DVS-2921: refactor COS/DVS hook for worker rebuilds
We want COS specific logic to be in the COS product stream to simplify
release coordination.  Arguably, all of the logic here was COS specific.
Some of the discovery commands were not COS specific, but we can think
of these as our interface to CSM.  All the discovery commands we ran
from the master node can be run from a worker as well, so we'll execute
a COS controlled file on the worker in question and allow everything to
take place there, with an implied value for TARGET_NCN.