CASMINST-5936: Use BOS v2 in prepare-images stage
For consistency with the managed-nodes-rollout stage, use BOS v2 when
creating BOS session templates in the prepare-images stage. Although
only the prepare-managed-images operation within that stage is currently
expected to create BOS session templates, technically the
prepare-management-images operation could create BOS session templates
if they were defined in the management bootprep file, so pass the
`--bos-version v2` option there as well.

This change is not necessary in the `update-cfs-config` stage because
the two operations in that stage pass only `--limit configurations` to
`sat bootprep run`, which means BOS session templates will not be
created.

Test Description:
Tested in combination with CASMINST-5885 on frigg by uploading the
modified Argo workflow templates to K8s and running the `prepare-images`
stage with simple bootprep files for managed and management nodes.