CASMINST-5885: Use INFO prefix on messages for skipped bootprep stages
Use the "INFO " prefix in the messages printed by the operations in the
"update-cfs-config" and "prepare-images" stages when no bootprep file is
provided to the operations in those stages.

This will ensure these messages are displayed to the end user at INFO
level.

Test Description:
Tested in combination with CASMINST-5936 on frigg by running the
update-cfs-config and prepare-images stages without passing in bootprep
files for either management or managed nodes and confirmed that the
INFO-level messages were logged to stdout by the CLI.

(cherry picked from commit 19e7f81653779ebfd74ba4bc334886215a1672f0)