CASMTRIAGE-5011 Respective cray-site-init install
The cray-site-init RPM is included in the tarball for multiple distros
and the `find` command is returning more than 1 RPM. The variable isn't
setup to contain more than one package, and breaks when it's passed to
the `rpm -Uvh` command.

This replaces the `find` with a `zypper in` command, and resolves the
respective distro to pull `cray-site-init` from.